requirements.txt
httpx
beautifulsoup4
lxml
tldextract
fastapi
uvicorn
jinja2
python-multipart
playwright
pydantic-settings
requests
PyYAML
PyJWT
fpdf2
slowapi
prometheus-fastapi-instrumentator
sentry-sdk
secure
redis
pydantic[email]
psutil
paramiko
sqlalchemy
alembic
google-api-python-client
google-auth
google-auth-httplib2
openpyxl
pandas